Add a simple CLI to manage Messaging Services (#54)
I intentionally kept this very simple because the plan is to
move messaging service management into the app itself. My plan
is to use this to update the service used for local development
with my ngrok url (the free version doesn't let you claim a url
for long-term use), and to provision services for production.