Don't pass statusCallback when using a messaging services (#66)
This is not needed when using Messaging Services. I'm not sure what Twilio picks when you pass both messagingServiceSid and statusCallback. Even if it picks statusCallback it doesn't make sense to keep it since any change to the base url will require us to update every Messaging Service's inboundRequestUrl. I'm keeping the env var around because it might makes sense to use it when provisioning Messaging Services through the ui.