Fix message review, update .env.example (#124)
Thinky strikes again! The tags table was defined correctly in the knex migration but not in the thinky model. This was broken for us but was probably working for the Working Families Party because initial database creation only runs the thinky createTables() function and assumes it never has to run any existing migration.

Fixes #25