Clean up data table styles (#127)
I know next to nothing about CSS, but I noticed that overflow: "scroll" was adding some ugly shadows to the cells in the data tables and didn't actually make anything more legible (it seems like textOverflow: "ellipsis" truncates the string regardless). This is a first pass and I think it makes the Message Review and People screens look pretty good.