Clean up opt outs (#134)
While the caching strategy for opt-outs was pretty good, it wasn't used consistently (you'll notice one place in the diff where we don't read through the cache) and you just really don't want to fuck around with opt outs. This is one place where I think it makes sense to do additional calls to the database to get the durability guarantees.

This also begins the effort of bringing order to the data access layer of the application by creating a db directory and adding a package for our opt out queries. This is a lightweight implementation of the repository pattern. I mostly just want data access in one place and to have a consistent way to thread transactions through. I'll document it in a separate PR.