Add external_id_type and state_code as top-level fields, optimize sync uploads (#141)
This change is probably not very useful for the community outside of political campaigns, but it is good for VAN users like us because we have multiple id types (MyC and MyV) and want to track state codes for all contacts. We also treat zip as an ordinary custom field because this fork does not appear to derive texting hours from zipcode anyway (others do).

In modifying the code I noticed we could avoid doing a bunch of work in CSV uploads when they are being processed with JOBS_SAME_PROCESS, so I threw in a little hack to speed it up.