Track canned response usage on the message table (#171)
I initially considered generalizing the question_response table, but it would have been a large refactor with few benefits, especially since we might want to track usage of canned responses that are not linked to survey questions. The main drawback of adding this to message is that it will make the query on the campaign stats page more expensive than the current question_response query, but I think we can deal with that in other ways if it becomes a problem.

I also deleted the server-side bulkSendMessage mutation since this is illegal in the USA and I don't want to have to updated it as we change the message sending api (although, fwiw, this particular change wouldn't have affected it).