Texter refactor: first pass (#288)
This is the first pass at a major refactor of the texting experience. It works but needs to be optimized and cleaned up in subsequent commits.

Conceptually we are starting to split the "Initial Send" and "Conversation" texting experiences. These haven't been split into separate components yet, but doing so in future commits will unlock a ton of optimizations,

A quick summary of what this changes:
- Adds sidebars matching the conversation texter to the initial send texter
- The AssignmentSummary card now has three sections: Send, Conversations, and Skipped
- The Conversation sidebar functions properly in the Conversation Texter view
- Arrows and message counters removed from the bottom nav, instead the initial send experience has a count down of messages to send (totally un-styled right now) 
- In dynamic assign mode, texters can request batches up to a certain size determined on the server
- Removed client-side caching of contacts, instead we only fetch a small amount of contact data from the server and request the full contact when we navigate to it. This works well for the conversation view but should be optimized for the initial send
- Chatting with a contact is more interactive because the AssignmentTexterContact view polls for updates to the conversation
- Fixed a bunch of small navigation and data fetching bugs
- Removes bulk send functionality


In the second iteration of this, I'd like to take the three components that make up the texting experience (`TexterTodo`, `AssignmentTexter`, and `AssignmentTexterContact`) and refactor them into 3 or 4 components: `InitialMessageTexter`, `ConversationTexter`, and either a generic `AssignmentTexterContact` component, or specific `InitialMessageTexterContact` and `ConversationTexterContact` components. It would also be nice to split the contact side bar out of the AssignmentTexterContact so we can refresh it separately from the current contact's data (the way we do this now is a little hacky and leads to a full screen re-render).