Upgrade Apollo Client (#294)
We were using an incredibly old version of Apollo that had a lot of problems. In particular, the client side cache didn't really work, so we were having to manually refresh in a lot of places where Apollo should have been picking up the changes automatically from the mutation results.

This gets us on version 1 of apollo, which is still like 2 years old but it not 4 years old like the version 0.4 we were using. 

The key changes are:
- `forceFetch: true` is now `fetchPolicy: "network-only"`
- Passing around `refetch` functions causes errors sometimes. Fortunately, you almost never need to manually refetch data.
- The data returned from the apollo cache is immutable, so I had to fix some places where we were calling `sort` on immutable data.
- There are some small differences in the calls to construct the apollo client
- The arguments and loading states using in `load-data.jsx` changed a bit, so I reworked that HOC to work correctly. I also added a trace mode to aid in future debugging of this component.